+++ libcoveocds/schema.py
@@ -266,7 +266,7 @@ class SchemaOCDS(SchemaJsonMixin):
     # less indirect to set the $ref to the file written by create_extended_schema_file() in this method rather than via
     # the resolver.
     #
-    # If using jsonschema>=4.18 and a lib-cove fork, self.registry is used to achieve this.
+    # If using jsonschema>=4.18 and lib-cove#123, self.registry is used to achieve this.
     #
     # For reference, the old code and original commit:
     # https://github.com/open-contracting/lib-cove-ocds/blob/19ed9b3f0e392e9341206c5296d79c4bcc6f1206/libcoveocds/schema.py#L196-L235  # noqa: E501