+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## 0.12.6 (2023-07-12)
+
+### Fixed
+
+- Include the extension URL in API context.
+
 ## 0.12.5 (2023-07-11)
 
-## Fixed
+### Fixed
 
 - Use the correct package schema in documentation URL fragments.
 - Use `JsonRef` proxies to determine the URL fragments for definitions' fields. (Regression in 0.12.4.)
@@ -16,7 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## 0.12.4 (2023-07-11)
 
-## Fixed
+### Fixed
 
 - Eliminate `JsonRef` proxies to avoid `AttributeError: 'JsonRef' object has no attribute 'get'` exceptions.